feat: added balance context provider
